Extract duplicated sanitize_for_yaml and study storage setup to optimisation/utils.py
Both run_optimization.py and run_sensitivity_analysis.py contained identical
sanitize_for_yaml() definitions and study storage setup logic. Consolidate
into a shared optimisation/utils.py module and import from there.

Closes #58
